==> scripts/mouse_CAR/scripts/03_FMT_alpha.Rmd <==
---
title: "FMT experiment"
author: "Anqi Dai"
date: '`r Sys.Date()`'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr)
library(vdbR)
connect_database()
get_table_from_database('shotgun_lookup_ad')
get_table_from_database('asv_annotation_blast_ag')

```

```{r}
FMT <- read_csv('../data/01_FMT_pheno.csv') %>% 
  mutate(Condition = str_replace(Condition, ' cells','')) %>% 
  mutate(condition_simple = case_when(
    Condition == '1st Collection (2 days) post-CAR T' ~ 'Week1 post-CAR',
    Condition == '2nd Week post-CAR T' ~ 'Week2 post-CAR',
    Condition == '3rd Week post-CAR T' ~ 'Week3 post-CAR',
    Condition == '4th Week post-CAR T' ~ 'Week4 post-CAR',
    Condition == '5th Week post-CAR T' ~ 'Week5 post-CAR',
    TRUE ~ Condition
  ))
```

# alpha diversity

```{r}
order_level <- FMT %>% 
  distinct(condition_simple, Date_Collected) %>% 
  arrange(Date_Collected) %>% 
  pull(condition_simple)

color_key <- FMT %>% 
  distinct(Group) %>% 
  mutate(color = c('#42B440','#EC0000')) %>% 
  deframe()

FMT %>% 
  #arrange(Date_Collected) %>% 
  mutate(condition_simple = factor(condition_simple, levels = order_level)) %>% 
  ggboxplot(x = 'Group', y = 'alpha', add = 'jitter', add.params = list( size = 3),color = 'Group',
            xlab = '', ylab = 'alpha diversity') +
  scale_color_manual(values = color_key) +
  facet_grid(. ~ condition_simple, scales = 'free') +
  theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  stat_compare_means(comparisons= list(c('Responder', 'Non-Responder')),
										label= "p.format",
										method= 'wilcox.test',
										correct=FALSE)

ggsave('../data/03_PH_alpha_FMT.pdf', width = 8, height = 6)
```

# Ruminococcus

```{r}
taxa <- read_csv('../data/01_all_metaphlan.csv') %>% 
  filter(sampleid %in% FMT$sampleid)

all <- read_csv('../data/01_all_metaphlan.csv')

rumino <- taxa %>% filter(str_detect(clade_name, 'Rumino'))

# g__Blautia|s__Ruminococcus_gnavus and g__Blautia|s__Ruminococcus_torques
# use the family level
rumino_fam <- taxa %>% 
  filter(str_detect(clade_name, 'f__Ruminococcaceae')) %>% 
  group_by(sampleid) %>% 
  summarise(rumino_family = sum(relative_abundance)) %>% 
  right_join(FMT, by = "sampleid") %>% 
  mutate(rumino_family = if_else(is.na(rumino_family), 0, rumino_family))
```

```{r}
rumino_fam %>% 
  #arrange(Date_Collected) %>% 
  mutate(condition_simple = factor(condition_simple, levels = order_level)) %>% 
  ggboxplot(x = 'Group', y = 'rumino_family', add = 'jitter', add.params = list( size = 3),color = 'Group',
            xlab = '', ylab = 'Ruminococcaceae relative abundance') +
  scale_color_manual(values = color_key) +
  facet_grid(. ~ condition_simple, scales = 'free') +
  theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  stat_compare_means(comparisons= list(c('Responder', 'Non-Responder')),
										label= "p.format",
										method= 'wilcox.test',
										correct=FALSE)

ggsave('../data/03_Ruminococcaceae_FMT.pdf', width = 8, height = 6)
```
```{r}
# the above plot but for g__Blautia|s__Ruminococcus_gnavus only
bromii <- taxa %>% 
  filter(str_detect(clade_name, 's__Ruminococcus_bromii')) %>% 
  right_join(FMT, by = "sampleid") %>% 
  mutate(relative_abundance = if_else(is.na(relative_abundance), 0, relative_abundance))

bromii %>% 
  filter(!Condition %in% c('Post-FMT 2','5th Week post-CAR T','4th Week post-CAR T')) %>% 
  #arrange(Date_Collected) %>% 
  mutate(condition_simple = factor(condition_simple, levels = order_level)) %>% 
  ggboxplot(x = 'Group', y = 'relative_abundance', add = 'jitter', add.params = list( size = 3),color = 'Group',
            xlab = '', ylab = 'Ruminococcus_bromii relative abundance') +
  scale_color_manual(values = color_key) +
  facet_grid(. ~ condition_simple, scales = 'free') +
  theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  stat_compare_means(comparisons= list(c('Responder', 'Non-Responder')),
										label= "p.format",
										method= 'wilcox.test',
										correct=FALSE)

ggsave('../data/03_Ruminococcus_bromii_FMT.pdf', width = 8, height = 6)
```

```{r}
# bacterial composition throughout time for responder vs not responder
genera <- all %>% 
  #filter(!str_detect(clade_name, '^k__Viruses')) %>% 
  mutate(genus = str_replace(clade_name, '\\|s__.+$','')) %>% 
  group_by(sampleid, genus) %>% 
  summarise(relab = sum(relative_abundance)) %>% 
  ungroup() %>% 
  mutate(ge = str_extract(genus, 'g__.+$')) %>%
  left_join(FMT, by = "sampleid") %>% 
  mutate(condition_simple = factor(condition_simple, levels = order_level)) %>% 
  arrange(Date_Collected, Group, Mouse_num) %>% 
  mutate(sid = str_glue('Condition {condition_simple}: Group {Group}: Mouse {Mouse_num}: {sampleid}')) 

genera %>% distinct(Mouse_num, Group, Condition)
colnames(genera)

genera %>% 
  ungroup() %>% 
  distinct(genus) %>% nrow

# find the genera > 0.1% in more than 10% samples
#   left 
keepgenera <- genera %>% 
  filter(relab > 0.001) %>% 
  count(genus) %>% 
  filter(n > floor(nrow(FMT) * 0.1)) %>% 
  pull(genus)

keepgenera <- genera %>% distinct(ge)
```

```{r}
#to find colors for these features
library(randomcoloR)
set.seed(1)
n <- length(keepgenera)
palette <- distinctColorPalette(n)
pie(rep(1, n), col=palette)

# output this palette otherwise another random generated ...

bar_palette <- genera %>%
  filter(genus %in% keepgenera) %>%
  distinct(ge) %>%
  arrange(ge) %>%
  mutate(color = palette)

bar_palette %>% write_csv('../data/03_genus_bar_palette.csv')

bar_plot_pal <- deframe(bar_palette)

bar_plot_pal <- read_csv('../data/03_genus_bar_palette.csv') %>% deframe
```


```{r}
# all samples
genera %>% 
  filter(genus %in% keepgenera) %>% 
  ggbarplot(x = 'sid', y = 'relab', fill = 'ge', color = 'white' , xlab = '', ylab = 'Relative abundance') +
  #facet_wrap(Day ~ fiber, scales = 'free') +
  scale_fill_manual(values = bar_plot_pal) +
  theme( axis.text.x = element_text(angle=90, hjust=0, size = 6.5),
         legend.position = 'bottom') 
ggsave('../data/03_bar_all.pdf', width = 14, height = 7)
```
```{r}
pheno <- read_csv('../data/01_pheno_with_alpha.csv')
```

```{r}
# the few samples 
few <- all %>% 
  filter(sampleid %in% c('PH18','PH19','PH20','PH21','PH22','PH23', 'PH24', 'PH25', 'PH26', 'PH131')) %>% 
  mutate(genus = str_replace(clade_name, '\\|s__.+$','')) %>% 
  group_by(sampleid, genus) %>% 
  summarise(relab = sum(relative_abundance)) %>% 
  ungroup() %>% 
  mutate(ge = str_extract(genus, 'g__.+$')) %>%
  left_join(FMT, by = "sampleid") %>% 
  mutate(condition_simple = factor(condition_simple, levels = order_level)) %>% 
  arrange(Date_Collected, Group, Mouse_num) %>% 
  mutate(sid = str_glue('Condition {condition_simple}: Group {Group}: Mouse {Mouse_num}: {sampleid}')) 

few %>% 
  filter(genus %in% keepgenera) %>% 
  ggbarplot(x = 'sid', y = 'relab', fill = 'ge', color = 'white' , xlab = '', ylab = 'Relative abundance') +
  #facet_wrap(Day ~ fiber, scales = 'free') +
  scale_fill_manual(values = bar_plot_pal) +
  theme( axis.text.x = element_text(angle=90, hjust=0, size = 6.5),
         legend.position = 'bottom') 

ggsave('../data/03_bar_few.pdf', width = 14, height = 7)
```

# to find the shared genus and percentage of between the slurry ad the mice samples

```{r}
slurry_responder <- c('PH131')
slurry_df <- genera %>% 
      filter(sampleid == slurry_responder)
responder_A20 <- FMT %>% 
  filter(Group == 'Responder' & Condition == 'Pre-A20') %>% 
  pull(sampleid)


# to find the shared genus and percentage of between the slurry and the pre-A20 samples
shared <- responder_A20 %>% 
  set_names(responder_A20) %>% 
  map(function(samp){
    A20_df = genera %>% filter(sampleid == samp)
    shared_genera = intersect(A20_df$ge, slurry_df$ge)
    relab_in_A20 = A20_df %>% 
      filter(ge %in% shared_genera) %>% group_by(sampleid) %>% summarise(shared_relab = sum(relab))
    relab_in_slurry = slurry_df %>% 
      filter(ge %in% shared_genera) %>% group_by(sampleid) %>% summarise(shared_relab = sum(relab))
    res = tibble(
      shared_genus = paste(shared_genera,collapse = ', '),
      relab_in_A20 = relab_in_A20$shared_relab,
      relab_in_slurry = relab_in_slurry$shared_relab
    )
  }) %>% bind_rows(.id = 'sampleid')

shared %>% write_csv('../data/03_shared_genera_between_PH131_and_A20_responder.csv')
```

```{r}
these <- FMT %>% 
  select(sampleid,Group, Condition ) %>% 
  add_row(sampleid = 'PH131', Group = 'Responder', Condition = 'slurry') %>% 
  filter(sampleid %in% c(responder_A20, 'PH131'))


# find the spp > 0.01% in more than 20% samples
keep_spp <- all %>% 
  filter(sampleid %in% these$sampleid) %>% 
  filter(relative_abundance > 0.0001) %>% 
  count(clade_name) %>% 
  filter(n > floor(nrow(these) * 0.2)) %>% 
  pull(clade_name)

# calculate the Bray curtis beta diversity 
cts_fil <- all %>% 
  filter(sampleid %in% these$sampleid) %>% 
  filter(clade_name %in% keep_spp) %>% 
  select(sampleid, clade_name,relative_abundance ) %>% 
  spread(key = 'clade_name', value = 'relative_abundance', fill = 0) %>% 
  column_to_rownames('sampleid')
dist_ <- vegan::vegdist(cts_fil, method = 'bray')
eigen <- cmdscale(dist_, eig = T)$eig
percent_var <- signif(eigen/sum(eigen), 3)*100
bc <- cmdscale(dist_, k = 2)
beta_df <- bc %>%
  as.data.frame() %>%
  rownames_to_column('sampleid')  %>% 
  full_join(these)

pcoa_all <- beta_df %>% 
  ggscatter(x = 'V1', y = 'V2', color =  'Condition', palette = 'lancet' ) +
  labs(title = 'Responders') +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  theme(aspect.ratio=1, legend.position = 'right')

ggsave(
  '../data/03_pcoa_10samples.pdf', width = 5, height = 4
)
```


```{r}
# the paired boxplot
shared %>% 
  gather('grp', 'relab', relab_in_A20:relab_in_slurry) %>%
  mutate(relab = round(relab, 2)) %>% 
  mutate(grp = factor(grp, levels = c('relab_in_slurry','relab_in_A20'))) %>% 
  ggbarplot(x= 'sampleid', y = 'relab',fill = "grp", color = "grp", palette = "Paired",
  label = TRUE,ylab = 'relative abundance',
  position = position_dodge(0.8))
ggsave('../data/03_shared_genera.pdf', width = 6, height = 5)
 
```
